moved pastebin into its own utility class
