+++ commands/help.js
@@ -1,9 +1,8 @@
 'use strict';
 
-const request = require('request');
 const runtime = require('../utils/Runtime');
 const Log = require('../utils/Log');
-const apiKey = 'c768c168d9bc885941e2e67e9cd8ad82';
+const Pastebin = require('../utils/Pastebin');
 
 module.exports = [{
     types: ['startup'],
@@ -22,24 +21,7 @@ module.exports = [{
             return;
         }
 
-        // Post the documentation to pastebin
-        Log.log('[help] Creating a new post on pastebin');
-        let requestOpts = {
-            url: 'http://pastebin.com/api/api_post.php',
-            form : {
-                api_dev_key: apiKey,
-                api_option: 'paste',
-                api_paste_code : newDocumentation,
-                api_paste_name: `${runtime.credentials.username} - LCTV Bot Help Documentation`
-            }
-        };
-        request.post( requestOpts, ( err, response, body ) => {
-            if ( err ) {
-                console.log( 'Error creating TextUploader post: ' + err );
-            }
-
-            Log.log('[help] Pastebin post created', body);
-
+        Pastebin.createPaste( `${runtime.credentials.username} - LCTV Bot Help Documentation`, newDocumentation, (body) => {
             helpObject.link = body;
             helpObject.documentation = newDocumentation;
             runtime.brain.set( 'help', helpObject );
@@ -47,7 +29,7 @@ module.exports = [{
     }
 }, {
 	name: '!help',
-	help: 'Lists the availabe commands.',
+	help: 'Shows the documentation for core, plugins, and custom commands.',
     types: ['message'],
     regex: /^(!|\/)(help|commands)$/,
     action: function( chat, stanza ) {
